Crawl_Ielts_Test/requirements.txt
Scrapy>=2.11.0,<3.0
beautifulsoup4>=4.12.0
pymongo>=4.6.0
python-slugify>=8.0.0

